refactor(#169): remove unused method
