--- src/comments/dto/get-all-comment-dto.ts
@@ -1,12 +1,0 @@
-export class CommentResponseDTO {
-  id: number;
-  content: string;
-  commentOwner: boolean;
-  user: {
-    name: string;
-    userId: number;
-    rank: number;
-    categoryId: number;
-    imageUrl: string;
-  };
-}
